requirements.txt
discord.py
configparser
discord_components
mcstatus
